+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.0.2
+
+**add**
+--
+- headline decoration
+- inline code
+
+**fix**
+--
+- bold style
+
+**code clean**
+--
+- moved each tag styles from style.dart to each elements files.
+
+
 ## 0.0.1
 
 yay! we were published super package!